+++ test/runtests.jl
@@ -45,6 +45,7 @@ if TEST_MODE in ("unit", "all")
 		include("./component/test_VaspIO.jl")
 		include("./component/test_vasp_to_extxyz.jl")
 		include("./component/test_poscar_to_toml.jl")
+		include("./component/test_outcar_to_embset.jl")
 		include("./component/test_thread_safety.jl")
 	end
 end